+++ src/main/java/com/example/demandForm/dto/request/FormOptionDto.java
@@ -1,4 +1,4 @@
-package com.example.demandForm.dto;
+package com.example.demandForm.dto.request;
 
 import jakarta.validation.constraints.Min;
 import lombok.AllArgsConstructor;
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class FormOptionDto {
 
     private Long optionId;
-    
+
     @Min(value = 1, message = "수량은 1개 이상이어야 합니다.")
     private int quantity;
 }
