+++ src/test/java/com/example/demandForm/service/DemandFormServiceTest.java
@@ -3,9 +3,9 @@ package com.example.demandForm.service;
 import com.example.common.exception.GlobalException;
 import com.example.demandForm.DemandTest;
 import com.example.demandForm.DemandTestBuilder;
-import com.example.demandForm.dto.CreateDemandFormRequestDto;
-import com.example.demandForm.dto.DemandFormNonMemberRequestDto;
-import com.example.demandForm.dto.DemandFormResponseDto;
+import com.example.demandForm.dto.request.CreateDemandFormRequestDto;
+import com.example.demandForm.dto.request.DemandFormNonMemberRequestDto;
+import com.example.demandForm.dto.response.DemandFormResponseDto;
 import com.example.demandForm.entity.DemandForm;
 import com.example.demandForm.entity.DemandOption;
 import com.example.demandForm.repository.DemandFormRepository;
